## Summary

- **Fix broken import paths** after service module reorganization
(power/, hardware/, sync/, versioning/, scheduler/)
- **Improve runtime safety** with null checks, input validation, type
casts, and SQLAlchemy `Mapped[]` annotations across 46 backend files
- **Fix sync schedule management** — weekday mapping, edit/delete,
device name resolution
- **Security audit & static analysis docs** added

## Changelog

### Added
- Mobile QR dialog auto-close on successful registration

### Fixed
- Import paths corrected for reorganized service modules
- Runtime crashes from wrong attribute access (RAID, audit)
- Token validation rejects missing `sub` claim
- Sleep/service endpoints accept empty request bodies
- Relative time formatter for future dates
- Deprecated patterns and name collisions removed

### Documentation
- Security audit report (2026-03-16)
- Static analysis report (ruff + mypy)

## Test plan
- [ ] Backend starts without ImportError
- [ ] Sync schedule CRUD works (create, edit, delete)
- [ ] Mobile QR pairing flow completes
- [ ] Admin VCL stats page loads
- [ ] Sleep mode endpoints respond correctly with empty body

+++ backend/app/schemas/mobile.py
@@ -200,7 +200,7 @@ class MobilePowerSummary(BaseModel):
 
     # CPU power profile
     power_profile: Optional[str] = None
-    power_profile_frequency_mhz: Optional[int] = None
+    power_profile_frequency_mhz: Optional[float] = None
     auto_scaling_enabled: bool = False
     active_demands_count: int = 0
